I applied more security on the cms to prevent user apart from admin to delete data

+++ cms/admin/INCLUDES/view_all_users.php
@@ -2,8 +2,7 @@
                             <thead>
                                 <tr>
                                     <th>Id</th>
-                                    <th>Username</th>
-                                    <th>Password</th>
+                                    <th>Username</th>                                    
                                     <th>Firstname</th>
                                     <th>Lastname</th>
                                     <th>Email</th>
@@ -25,8 +24,7 @@
     while($row = mysqli_fetch_assoc($select_users))
     {
         $user_id = $row['user_id'];
-        $username = $row['username'];
-        $user_password = $row['user_password'];
+        $username = $row['username'];        
         $user_firstname = $row['user_firstname'];
         $user_lastname = $row['user_lastname'];
         $user_email = $row['user_email'];        
@@ -37,8 +35,7 @@
 
         echo "<tr>";
         echo "<td>{$user_id}</td>";
-        echo "<td>{$username}</td>";
-        echo "<td>{$user_password}</td>";
+        echo "<td>{$username}</td>";        
         echo "<td>{$user_firstname}</td>";
         echo "<td>{$user_lastname}</td>";
         echo "<td>{$user_email}</td>";
@@ -97,12 +94,20 @@ if(isset($_GET['change_to_sub']))
 
 if(isset($_GET['delete']))
 {
-    $the_user_id = $_GET['delete'];
+    if(isset($_SESSION['user_role']))
+    {
+        if($_SESSION['user_role'] == 'Admin')
+        {
+            $the_user_id = mysqli_real_escape_string($connection, $_GET['delete']);
     
-    $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
+            $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
+
+            $delete_query = mysqli_query($connection, $query);
+            header("Location: users.php");
+        }        
+    }
     
-    $delete_query = mysqli_query($connection, $query);
-    header("Location: users.php");
+    
 }
